LPS-166857 Moves implementations to layout-impl
