+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/exporter/PortletPreferencesPortletConfigurationExporterImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.exporter;
+package com.liferay.layout.internal.exporter;
 
 import com.liferay.layout.exporter.PortletPreferencesPortletConfigurationExporter;
 import com.liferay.petra.string.StringPool;
